+++ packages/checkbox/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.2](https://github.com/Availity/element/compare/@availity/mui-checkbox@0.1.1...@availity/mui-checkbox@0.1.2) (2023-10-17)
+
 ## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-checkbox@0.1.0...@availity/mui-checkbox@0.1.1) (2023-10-04)
 
 ## 0.1.0 (2023-10-04)