+++ packages/theme/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.21.0](https://github.com/Availity/element/compare/@availity/theme@0.20.0...@availity/theme@0.21.0) (2023-10-20)
+
+
+### Features
+
+* **mui-form-utils:** add FormHelperText and Select ([0ff9d8e](https://github.com/Availity/element/commit/0ff9d8e06d0baac263ce1869304d3b91fcb5cb15))
+
+## [0.20.0](https://github.com/Availity/element/compare/@availity/theme@0.19.4...@availity/theme@0.20.0) (2023-10-17)
+
+
+### Features
+
+* **mui-card:** add legacy theme ([ff7e238](https://github.com/Availity/element/commit/ff7e2381c686f27cfc6ada02b1a4c315a28c90a5))
+
+## [0.19.4](https://github.com/Availity/element/compare/@availity/theme@0.19.3...@availity/theme@0.19.4) (2023-10-17)
+
+### Bug Fixes
+
+- **mui-menu:** selected active state styling ([94a80df](https://github.com/Availity/element/commit/94a80dfbb87a73be2e4c2f02afd8bbd797f5e784))
+
 ## [0.19.3](https://github.com/Availity/element/compare/@availity/theme@0.19.2...@availity/theme@0.19.3) (2023-10-17)
 
 ## [0.19.2](https://github.com/Availity/element/compare/@availity/theme@0.19.1...@availity/theme@0.19.2) (2023-10-13)